requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cffi==1.14.3              # via cryptography
click==7.1.2              # via flask
cryptography==3.2.1       # via VanasTokenManager (setup.py)
flask==1.1.2              # via VanasTokenManager (setup.py)
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via VanasTokenManager (setup.py)
six==1.15.0               # via cryptography
werkzeug==1.0.1           # via VanasTokenManager (setup.py), flask

git+https://github.com/liuqi0725/Vanas-PyHelper#egg=VanasPyHelper
# VanasPyHelper 相关包
chardet==3.0.4
pycrypto==2.6.1
Pillow==8.0.1
pyyaml==5.3.1
konfig
werkzeug

